Polish backoff and retry support
This revises commit 15dd320b95aa404c9542e029588c7aefef5c37ff.

See gh-34529
See gh-35110

+++ spring-core/src/main/java/org/springframework/util/backoff/ExponentialBackOff.java
@@ -90,7 +90,7 @@ public class ExponentialBackOff implements BackOff {
 	 * The default maximum attempts: unlimited.
 	 * @since 6.1
 	 */
-	public static final int DEFAULT_MAX_ATTEMPTS = Integer.MAX_VALUE;
+	public static final long DEFAULT_MAX_ATTEMPTS = Long.MAX_VALUE;
 
 
 	private long initialInterval = DEFAULT_INITIAL_INTERVAL;
